# dtl/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DTL API
  description: |
    API for DTL service (quote requests, appointment requests, push, settings, admin CRUD).
    See api.md for narrative documentation.
    Admin and settings endpoints are implemented as Next.js API routes; they use in-memory or
    localStorage-backed stores (no external DB yet).
  version: 1.0.0
servers:
  - url: /api
    description: Same-origin or NEXT_PUBLIC_API_URL base
...
security:
  - bearerAuth: []
paths:
...
  /users/{id}:
    get:
      tags: [Admin – Users]
      summary: Get user by id
      operationId: getUserById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      tags: [Admin – Users]
      summary: Update user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    delete:
      tags: [Admin – Users]
      summary: Delete user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /quote-requests:
    get:
      tags: [Admin – Quote requests]
      summary: List quote requests
      description: Implemented via Next.js route; uses in-memory store.
      operationId: getQuoteRequests
      responses:
        "200":
          description: List of quote requests
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/QuoteRequest"
    post:
      tags: [Frontend – Quote requests]
      summary: Create a quote request
      operationId: createQuoteRequest
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestCreate"
      responses:
        "201":
          description: Created quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "400":
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /settings/feature-flags:
    get:
      tags: [Frontend – Feature flags]
      summary: Get feature flags
      description: Returns object merged with client defaults. When showUnavailableSlots is true, booking UI should show unavailable slots (e.g. greyed out) to normal users.
      operationId: getFeatureFlags
      security: []
      responses:
        "200":
          description: Feature flags object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlags"
    put:
      tags: [Admin – Feature flags]
      summary: Update feature flags
      operationId: adminPutFeatureFlags
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FeatureFlags"
      responses:
        "200":
          description: Updated feature flags
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeatureFlags"
  /settings/business:
    get:
      tags: [Frontend – Business settings]
      summary: Get business settings
      operationId: getBusinessSettings
      security: []
      responses:
        "200":
          description: Business settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"
    put:
      tags: [Admin – Business settings]
      summary: Update business settings
      operationId: adminPutBusinessSettings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BusinessSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessSettings"
...
  /testimonials:
    get:
      tags: [Admin – Testimonials]
      summary: List testimonials
      operationId: adminGetTestimonials
      security: []
      responses:
        "200":
          description: List of testimonials
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/TestimonialItem"
    post:
      tags: [Admin – Testimonials]
      summary: Add testimonial
      operationId: adminAddTestimonial
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialItemCreate"
      responses:
        "201":
          description: Created testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestimonialItem"
  /testimonials/{id}:
    patch:
      tags: [Admin – Testimonials]
      summary: Update testimonial
      operationId: adminUpdateTestimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestimonialItemUpdate"
      responses:
        "200":
          description: Updated testimonial
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestimonialItem"
        "404":
          description: Not found
    delete:
      tags: [Admin – Testimonials]
      summary: Delete testimonial
      operationId: adminDeleteTestimonial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /settings/working-hours:
    get:
      tags: [Admin – Working hours]
      summary: Get working hours (orar)
      operationId: adminGetWorkingHours
      security: []
      responses:
        "200":
          description: Working hours schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkingHoursSchedule"
    put:
      tags: [Admin – Working hours]
      summary: Update working hours
      operationId: adminPutWorkingHours
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkingHoursSchedule"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkingHoursSchedule"
  /settings/contact:
    get:
      tags: [Admin – Contact]
      summary: Get contact settings
      operationId: adminGetContact
      security: []
      responses:
        "200":
          description: Contact settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSettings"
    put:
      tags: [Admin – Contact]
      summary: Update contact settings
      operationId: adminPutContact
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSettings"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSettings"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
        createdAt:
          type: string
          format: date-time
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        carMake:
          type: string
        carModel:
          type: string
        carYear:
          type: string
        description:
          type: string
        chassis:
          type: string
        photoUrl:
          type: string
        status:
          type: string
    QuoteRequestCreate:
      type: object
      required:
        - name
        - phone
        - email
        - carMake
        - carModel
        - carYear
        - description
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        carMake:
          type: string
        carModel:
          type: string
        carYear:
          type: string
        description:
          type: string
        chassis:
          type: string
        photoUrl:
          type: string
...
    FeatureFlags:
      type: object
      description: All properties optional booleans; merged with client defaults.
      properties:
        tyreService:
          type: boolean
        carWash:
          type: boolean
        requestQuote:
          type: boolean
        generalServiceBooking:
          type: boolean
        tyreServiceBooking:
          type: boolean
        carWashBooking:
          type: boolean
        partsOrdering:
          type: boolean
        cardInstallmentPayment:
          type: boolean
        authentication:
          type: boolean
        tyreServiceVisible:
          type: boolean
        carWashVisible:
          type: boolean
        requestQuoteVisible:
          type: boolean
        generalServiceBookingVisible:
          type: boolean
        tyreServiceBookingVisible:
          type: boolean
        carWashBookingVisible:
          type: boolean
        partsOrderingVisible:
          type: boolean
        cardInstallmentPaymentVisible:
          type: boolean
        authenticationVisible:
          type: boolean
        showServicePrices:
          type: boolean
        showTyreServicePrices:
          type: boolean
        showCarWashPrices:
          type: boolean
        gallery:
          type: boolean
        testimonials:
          type: boolean
        showUnavailableSlots:
          type: boolean
          description: When true, booking UI shows unavailable slots to normal users (e.g. greyed out).
    BusinessSettings:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        logoUrl:
          type: string
        hours:
          type: object
          additionalProperties:
            type: string
...
    Error:
      type: object
      properties:
        error:
          type: string

    # Admin / backend entities
...
    WorkingHoursSchedule:
      type: object
      description: "Per-day schedule Mon-Sat (lib/working-hours.ts). Each day is either start/end or closed true."
      required: [days]
      properties:
        days:
          type: array
          minItems: 6
          maxItems: 6
          items:
            type: object
            description: "Either start/end (HH:mm) or closed true"
            properties:
              start:
                type: string
                description: HH:mm
              end:
                type: string
                description: HH:mm
              closed:
                type: boolean
    ContactSettings:
      type: object
      properties:
        companyName:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
...
    TestimonialItem:
      type: object
      properties:
        id:
          type: string
        author:
          type: string
        role:
          type: string
        text:
          type: string
        rating:
          type: number
        createdAt:
          type: string
          format: date-time
        visible:
          type: boolean
    TestimonialItemCreate:
      type: object
      properties:
        author:
          type: string
        role:
          type: string
        text:
          type: string
        rating:
          type: number
        visible:
          type: boolean
    TestimonialItemUpdate:
      type: object
      description: Partial update for testimonial.
      properties:
        author:
          type: string
        role:
          type: string
        text:
          type: string
        rating:
          type: number
        visible:
          type: boolean
...
    User:
      type: object
      description: Admin user (lib/users-store.ts, dummy-users.ts).
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Super Admin, Admin, Staff]
        active:
          type: boolean
        canManageUsers:
          type: boolean
        lastLogin:
          type: string
          format: date-time
...
    UserUpdate:
      type: object
      description: Partial update for user.
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [Admin, Staff]
        active:
          type: boolean
        canManageUsers:
          type: boolean